+++ x/btcstaking/keeper/btc_delegations.go
@@ -224,41 +224,28 @@ func (k Keeper) setBTCDelegation(ctx context.Context, btcDel *types.BTCDelegatio
 // validateMultiStakedFPs ensures all finality providers are known to Babylon and at least
 // one of them is a Babylon finality provider. It also checks whether the BTC stake is
 // multi-staked to FPs of consumer chains
-func (k Keeper) validateMultiStakedFPs(ctx context.Context, fpBTCPKs []bbn.BIP340PubKey) (bool, error) {
+func (k Keeper) validateMultiStakedFPs(ctx sdk.Context, fpBTCPKs []bbn.BIP340PubKey) error {
 	multiStakedToBabylon := false
-	multiStakedToConsumers := false
 
 	for i := range fpBTCPKs {
 		fpBTCPK := fpBTCPKs[i]
 
-		// find the fp and determine whether it's Babylon fp or consumer chain fp
-		if fp, err := k.GetFinalityProvider(ctx, fpBTCPK); err == nil {
-			// ensure the finality provider is not slashed
-			if fp.IsSlashed() {
-				return false, types.ErrFpAlreadySlashed
-			}
+		fp, err := k.GetFinalityProvider(ctx, fpBTCPK)
+		if err != nil {
+			return types.ErrFpNotFound.Wrapf("finality provider pk %s is not found: %v", fpBTCPK.MarshalHex(), err)
+		}
+		if fp.IsSlashed() {
+			return types.ErrFpAlreadySlashed
+		}
+		if fp.SecuresBabylonGenesis(ctx) {
 			multiStakedToBabylon = true
-			continue
-		} else if consumerID, err := k.BscKeeper.GetConsumerOfFinalityProvider(ctx, &fpBTCPK); err == nil {
-			fp, err := k.BscKeeper.GetConsumerFinalityProvider(ctx, consumerID, &fpBTCPK)
-			if err != nil {
-				return false, err
-			}
-			// ensure the finality provider is not slashed
-			if fp.IsSlashed() {
-				return false, types.ErrFpAlreadySlashed
-			}
-			multiStakedToConsumers = true
-			continue
-		} else {
-			return false, types.ErrFpNotFound.Wrapf("finality provider pk %s is not found", fpBTCPK.MarshalHex())
 		}
 	}
 	if !multiStakedToBabylon {
 		// a BTC delegation has to stake to at least 1 Babylon finality provider
-		return false, types.ErrNoBabylonFPRestaked
+		return types.ErrNoBabylonFPRestaked
 	}
-	return multiStakedToConsumers, nil
+	return nil
 }
 
 // multiStakedFPConsumerIDs returns the unique consumer IDs of non-Babylon finality providers
@@ -268,12 +255,12 @@ func (k Keeper) multiStakedFPConsumerIDs(ctx context.Context, fpBTCPKs []bbn.BIP
 
 	for i := range fpBTCPKs {
 		fpBTCPK := fpBTCPKs[i]
-		if _, err := k.GetFinalityProvider(ctx, fpBTCPK); err == nil {
-			continue
-		} else if consumerID, err := k.BscKeeper.GetConsumerOfFinalityProvider(ctx, &fpBTCPK); err == nil {
-			consumerIDMap[consumerID] = struct{}{}
-		} else {
-			return nil, types.ErrFpNotFound.Wrapf("finality provider pk %s is not found", fpBTCPK.MarshalHex())
+		fp, err := k.GetFinalityProvider(ctx, fpBTCPK)
+		if err != nil {
+			return nil, err
+		}
+		if !fp.SecuresBabylonGenesis(sdk.UnwrapSDKContext(ctx)) {
+			consumerIDMap[fp.BsnId] = struct{}{}
 		}
 	}