+++ x/btcstaking/keeper/genesis.go
@@ -17,7 +17,6 @@ import (
 
 // InitGenesis initializes the module's state from a provided genesis state.
 func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
 	// save all past params versions
 	for _, p := range gs.Params {
 		params := p
@@ -81,10 +80,6 @@ func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
 		}
 	}
 
-	if err := k.setBTCConsumerDelegators(sdkCtx, gs.BtcConsumerDelegators); err != nil {
-		return err
-	}
-
 	if err := k.setConsumerEvents(ctx, gs.ConsumerEvents); err != nil {
 		return err
 	}
@@ -119,11 +114,6 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		return nil, err
 	}
 
-	consumerBtcDels, err := k.btcDelegatorsWithKey(ctx, types.BTCConsumerDelegatorKey)
-	if err != nil {
-		return nil, err
-	}
-
 	return &types.GenesisState{
 		Params:                 k.GetAllParams(ctx),
 		FinalityProviders:      fps,
@@ -133,7 +123,6 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		Events:                 evts,
 		AllowedStakingTxHashes: txHashes,
 		LargestBtcReorg:        k.GetLargestBtcReorg(ctx),
-		BtcConsumerDelegators:  consumerBtcDels,
 		ConsumerEvents:         k.consumerEvents(ctx),
 	}, nil
 }
@@ -310,30 +299,6 @@ func (k Keeper) setEventIdx(
 		return err
 	}
 	store.Set(sdk.Uint64ToBigEndian(evt.Idx), bz)
-
-	return nil
-}
-
-// setBTCConsumerDelegators is a function to set the provided BTC
-// consumer delegators data. It thows an error if the provided FP
-// does not exist or there's an already
-// existing store entry for the provided data.
-// NOTE: this is used in InitGenesis only
-func (k Keeper) setBTCConsumerDelegators(ctx sdk.Context, btcDels []*types.BTCDelegator) error {
-	for _, btcDel := range btcDels {
-		if !k.BscKeeper.HasConsumerFinalityProvider(ctx, btcDel.FpBtcPk) {
-			return fmt.Errorf("finality provider not found. BTC pk: %s", btcDel.FpBtcPk.MarshalHex())
-		}
-
-		// get BTC delegation index under this finality provider
-		btcDelIndex := k.getBTCConsumerDelegatorDelegationIndex(ctx, btcDel.FpBtcPk, btcDel.DelBtcPk)
-		if btcDelIndex != nil {
-			return fmt.Errorf("delegation index already exists. FP BTC pk: %s, Delegator BTC pk: %s", btcDel.FpBtcPk.MarshalHex(), btcDel.DelBtcPk.MarshalHex())
-		}
-
-		// save the index
-		k.setBTCConsumerDelegatorDelegationIndex(ctx, btcDel.FpBtcPk, btcDel.DelBtcPk, btcDel.Idx)
-	}
 
 	return nil
 }