+++ x/btcstaking/keeper/genesis_test.go
@@ -167,7 +167,7 @@ func setupTest(t *testing.T) (sdk.Context, *helper.Helper, *types.GenesisState) 
 	k, ctx := h.App.BTCStakingKeeper, h.Ctx
 	numFps := 3
 
-	fps := datagen.CreateNFinalityProviders(r, t, h.FpPopContext(), numFps)
+	fps := datagen.CreateNFinalityProviders(r, t, h.FpPopContext(), "", numFps)
 	params := k.GetParams(ctx)
 
 	chainsHeight := make([]*types.BlockHeightBbnToBtc, 0)
@@ -180,7 +180,6 @@ func setupTest(t *testing.T) (sdk.Context, *helper.Helper, *types.GenesisState) 
 	events := make([]*types.EventIndex, 0)
 	btcDelegators := make([]*types.BTCDelegator, 0)
 	allowedStkTxHashes := make([]string, 0)
-	consumerBtcDelegators := make([]*types.BTCDelegator, 0)
 	consumerEvents := make([]*types.ConsumerEvent, 0)
 
 	blkHeight := uint64(r.Int63n(1000)) + math.MaxUint16
@@ -265,7 +264,6 @@ func setupTest(t *testing.T) (sdk.Context, *helper.Helper, *types.GenesisState) 
 		Events:                 events,
 		AllowedStakingTxHashes: allowedStkTxHashes,
 		LargestBtcReorg:        latestBtcReOrg,
-		BtcConsumerDelegators:  consumerBtcDelegators,
 		ConsumerEvents:         consumerEvents,
 	}
 	require.NoError(t, gs.Validate())