+++ x/btcstaking/types/finality_provider.go
@@ -1,0 +1,7 @@
+package types
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+func (fp *FinalityProvider) SecuresBabylonGenesis(ctx sdk.Context) bool {
+	return fp.BsnId == ctx.ChainID()
+}